api/requirements.txt
anyio==3.3.2
argon2-cffi==21.1.0
asgiref==3.4.1
Babel==2.9.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.1
cryptography==35.0.0
decorator==5.1.0
Django==3.1.13
django-classy-tags==2.0.0
django-cms==3.8.0
django-cors-headers==3.10.0
django-cursor-pagination==0.1.5
django-filer==2.0.2
django-formtools==2.3
django-js-asset==1.2.2
django-money==2.1
django-mptt==0.13.4
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.7.1
graphql-core==3.1.6
h11==0.12.0
html5lib==1.1
idna==3.2
importlib-metadata==4.8.1
phonenumberslite==8.12.34
Pillow==8.3.2
psycopg2-binary==2.9.1
py-moneyed==1.2
pycparser==2.20
Pygments==2.10.0
PyJWT==2.2.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.3
requests==2.26.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
strawberry-graphql==0.81.0
typing-extensions==3.10.0.2
Unidecode==1.1.2
urllib3==1.26.7
uvicorn==0.15.0
validators==0.18.2
webencodings==0.5.1
zipp==3.6.0
